requirements.txt
discord.py>=2.3.0
gspread>=5.12.0
google-auth>=2.23.0
google-cloud-vision>=3.5.0
google-api-python-client>=2.100.0
python-dotenv>=1.0.0
